Solve password hash when uploading video n Add user authority

+++ src/controllers/videoController.js
@@ -18,19 +18,28 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {user: {_id}} = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  console.log(typeof video.owner, typeof _id);  
+  if(String(video.owner) !== String(_id)){  // video 올린 user의 id(owner)와 로그인한 user의 id가 같지 않으면
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {user: {_id}} = req.session;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if(String(video.owner) !== String(_id)){  // video 올린 user의 id(owner)와 로그인한 user의 id가 같지 않으면
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -73,6 +82,14 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {user: {_id}} = req.session;
+  const video = await Video.findById(id);
+  if(!video){
+    return res.status(404).render("404", {pageTitle: "Video not found"});
+  }
+  if(String(video.owner) !== String(_id)){  // video 올린 user의 id(owner)와 로그인한 user의 id가 같지 않으면
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
